feat: add QR generator with batch downloads
